refactor: transfer classes of ongoing matches to other packages.
