+++ src/main/java/com/projects/tennisscoreboard/dto/match/ongoing/MatchCreateDto.java
@@ -1,4 +1,4 @@
-package com.projects.tennisscoreboard.dto;
+package com.projects.tennisscoreboard.dto.match.ongoing;
 
 import lombok.Builder;
