+++ src/main/java/com/projects/tennisscoreboard/dto/match/ongoing/MatchProgressDto.java
@@ -1,5 +1,7 @@
-package com.projects.tennisscoreboard.dto;
+package com.projects.tennisscoreboard.dto.match.ongoing;
 
+import com.projects.tennisscoreboard.dto.match.MatchState;
+import com.projects.tennisscoreboard.dto.match.ScoreDto;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
