+++ src/main/java/com/projects/tennisscoreboard/validator/impl/CreateMatchValidator.java
@@ -1,7 +1,7 @@
 package com.projects.tennisscoreboard.validator.impl;
 
 import com.projects.tennisscoreboard.Utils.PropertiesUtil;
-import com.projects.tennisscoreboard.dto.MatchCreateDto;
+import com.projects.tennisscoreboard.dto.match.ongoing.MatchCreateDto;
 import com.projects.tennisscoreboard.validator.ValidationError;
 import com.projects.tennisscoreboard.validator.ValidationResult;
 import com.projects.tennisscoreboard.validator.Validator;
